requirements.txt
gradio==3.50.2
python-dotenv==1.1.1
sentence-transformers==5.1.0
langchain==0.0.354
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
bs4==4.8.2
ebooklib==0.19
textract==1.6.5
tqdm==4.67.1
aiofiles==23.2.1
aiohttp==3.12.15
